shorten project
