+++ src/app/project/component.ts
@@ -49,8 +49,8 @@ import {EditProjectComponent} from './edit-project/edit-project.component';
 
 @Component({
   selector: 'km-project',
-  templateUrl: './project.component.html',
-  styleUrls: ['./project.component.scss'],
+  templateUrl: './template.html',
+  styleUrls: ['./style.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class ProjectComponent implements OnInit, OnChanges, OnDestroy {
