+++ src/app/project/routing.ts
@@ -10,17 +10,20 @@
 // limitations under the License.
 
 import {NgModule} from '@angular/core';
-import {SharedModule} from '../shared/shared.module';
-import {EditProjectComponent} from './edit-project/edit-project.component';
-import {ProjectRoutingModule} from './project-routing.module';
-import {ProjectComponent} from './project.component';
-import {DeleteProjectConfirmationComponent} from './delete-project/delete-project.component';
+import {RouterModule, Routes} from '@angular/router';
+import {AuthGuard} from '@core/services/auth/guard';
+import {ProjectComponent} from './component';
 
-const components: any[] = [ProjectComponent, EditProjectComponent, DeleteProjectConfirmationComponent];
+const routes: Routes = [
+  {
+    path: '',
+    component: ProjectComponent,
+    canActivate: [AuthGuard],
+  },
+];
 
 @NgModule({
-  imports: [SharedModule, ProjectRoutingModule],
-  declarations: [...components],
-  exports: [...components],
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule],
 })
-export class ProjectModule {}
+export class ProjectRoutingModule {}
